backend-issue-matcher/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
pymupdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
python-multipart
